+++ changelog.md
@@ -2,6 +2,8 @@
 ### Additions
 - Added the following config options:
   - `base_experience_cap`: Controls how much experience you need for one level of a tool
+  - `experience_multiplier`: Controls how much experience required changes per level
+  - `armor_experience_boost`: Controls how much more experience armor gets when you take damage
 
 ----------------------- 0.5.0 Changes -----------------------
 ### Additions